[master_025 -> test] Adding database autocomplete on startup. Changing the data type for ID from Guid -> long (+identity)

+++ Back/MishFit/MishFit/Services/MealsService.cs
@@ -1,5 +1,6 @@
 ﻿using MishFit.Contracts;
 using MishFit.Entities;
+using MishFit.Exceptions;
 using MishFit.Repositories;
 
 namespace MishFit.Services;
@@ -18,9 +19,19 @@ public class MealsService : IMealsService
         return await _repository.GetAllMealsAsync();
     }
 
-    public async Task<Meal> GetMealByIdAsync(Guid id)
+    public async Task<Meal> GetMealByIdAsync(long id)
     {
+        if (id <= 0)
+            throw new InvalidIncomingParameterException($"Meal id cannot be null.");
         return await _repository.GetMealByIdAsync(id);
+    }
+    
+    public async Task<List<Meal>> SearchMealByNameAsync(string name)
+    {
+        if (name.Trim()=="")
+            throw new InvalidIncomingParameterException($"Meal name cannot be null.");
+        
+        return await _repository.SearchMealByNameAsync(name.ToLower().Trim());
     }
 
     public async Task<Meal> CreateMealAsync(CreateMealContract contract)
@@ -33,8 +44,11 @@ public class MealsService : IMealsService
         return await _repository.UpdateMealAsync(contract);
     }
 
-    public async Task<Meal> DeleteMealByIdAsync(Guid id)
+    public async Task<Meal> DeleteMealByIdAsync(long id)
     {
+        if (id <= 0)
+            throw new InvalidIncomingParameterException($"Meal id cannot be null.");
+
         return await _repository.DeleteMealByIdAsync(id);
     }
 }
